User and home page changes
