+++ src/context/SidebarContext.tsx
@@ -8,7 +8,7 @@ type SidebarContextType = {
 const SidebarContext = createContext<SidebarContextType | undefined>(undefined);
 
 export const SidebarProvider = ({ children }: { children: React.ReactNode }) => {
-  const [selectedTab, setSelectedTab] = useState<string>('Users'); // default tab
+  const [selectedTab, setSelectedTab] = useState<string>('Home'); // default tab
 
   return (
     <SidebarContext.Provider value={{ selectedTab, setSelectedTab }}>
