--- src/features/Dashboard/Home.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-const HomePage = () => {
-    return (
-        <>
-            <h1>hello</h1>
-        </>
-    )
-}
-
-export default HomePage;
